audio: DEXED passe de 8 a 255 vrais patches d'usine DX7 (ROM1-ROM4)
Demande : 'recuperer un max de patch pour tout les moteur'. Trouvaille :
banks.h (deja vendored avec Synth_Dexed) contient 320 vrais patches DX7,
dont les 8 vraies banques ROM d'usine du Yamaha DX7 original (256
patches, les presets les plus connus du DX7) -- jamais exploite avant
ce soir, seuls 8 patches choisis a la main etaient utilises.

- 255 des 256 patches ROM extraits (noms pris du sysex lui-meme) vers
  az2_dexed_bank_data.h. 255 et PAS 256 : 0xFF/255 est deja le sentinel
  'pas d'override de patch par pas' (stepPatch/seqStepPatch) -- un
  vrai patch DEXED 255 aurait ete indiscernable de ce sentinel, le
  dernier patch de ROM4B ('EXPLOSION') est laisse de cote.
- Bug trouve et corrige au passage : enginePatchCount() retournait un
  uint8_t -- un compte de 256 (avant la decouverte du piege 0xFF)
  s'y serait tronque en 0, '% count' devenant une division par zero.
  Elargi en uint16_t, 3 sites d'appel ESP32 corriges au meme type.
- Test natif mis a jour (8 -> 255).

Teste sur le vrai materiel (piste isolee, patches 0/127/254 a la
suite) : 3 sons distincts confirmes sur toute la plage.

+++ src_esp32/az2_screen/main.cpp
@@ -3061,7 +3061,7 @@ void handleTeensyLine(const String &line) {
               const uint8_t nextEngine = static_cast<uint8_t>((trackEngine[t] + dir + az2::kEngineCount) % az2::kEngineCount);
               snprintf(msg, sizeof(msg), "ENGINE:%d:%d", t, nextEngine);
             } else {
-              const uint8_t count = az2::enginePatchCount(trackEngine[t]);
+              const uint16_t count = az2::enginePatchCount(trackEngine[t]);  // uint16_t depuis DEXED=256 patches (2026-09-18)
               const uint8_t nextPatch = static_cast<uint8_t>((trackPatch[t] + dir + count) % count);
               snprintf(msg, sizeof(msg), "PATCH:%d:%d", t, nextPatch);
             }
@@ -3126,7 +3126,7 @@ void handleTeensyLine(const String &line) {
                   break;
                 }
                 case 1: {
-                  const uint8_t count = az2::enginePatchCount(trackEngine[t]);
+                  const uint16_t count = az2::enginePatchCount(trackEngine[t]);  // uint16_t depuis DEXED=256 patches (2026-09-18)
                   const int base = (seqStepPatch[currentPattern][t][s] == 0xFF) ? trackPatch[t] : seqStepPatch[currentPattern][t][s];
                   const int newPatch = (base + dir + count) % count;
                   snprintf(msg, sizeof(msg), "INST:%d:%d:%d", t, s, newPatch);
@@ -4338,7 +4338,7 @@ void handleTouchDown(uint8_t slot, int16_t x, int16_t y) {
         const uint8_t nextEngine = static_cast<uint8_t>((trackEngine[track] + 1) % az2::kEngineCount);
         snprintf(msg, sizeof(msg), "ENGINE:%d:%d", track, nextEngine);
       } else {
-        const uint8_t count = az2::enginePatchCount(trackEngine[track]);
+        const uint16_t count = az2::enginePatchCount(trackEngine[track]);  // uint16_t depuis DEXED=256 patches (2026-09-18)
         const uint8_t nextPatch = static_cast<uint8_t>((trackPatch[track] + 1) % count);
         snprintf(msg, sizeof(msg), "PATCH:%d:%d", track, nextPatch);
       }
